fix(analytic):change UserService to client
change the UserService to UserServiceClient
