+++ src/main/java/faang/school/analytics/dto/user/UserDto.java
@@ -1,0 +1,10 @@
+package faang.school.analytics.dto.user;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class UserDto {
+    private long id;
+}
